Filstrukturen er ordnet. Dokumentation til samtlige funktioner er tilføjet/strømlignet og route_count og train_count er nu dynamisk genereret baseret på *.txt filerne
