+++ src/routes.h
@@ -1,15 +1,9 @@
 #ifndef P1_ROUTES_H
 #define P1_ROUTES_H
+
 #define DATA_SIZE 40
-#define ROUTE_COUNT 6
-#include <string.h>
-// Data taken from "https://www.raileurope.com/en"
-// For more info see "https://www.raileurope.com/blog/travel-green-calculating-your-carbon-savings"
 
-// When multiple routes for the same journey are available, the route with shortest travel time is selected.
-
-// All stations are of format "SOURCE_DESTINATION".
-
+// Struct that holds data for routes - aka. the template
 typedef struct route{
     int distance;
     int track_speed;
@@ -20,131 +14,37 @@ typedef struct route{
     char station_end[DATA_SIZE];
 } route;
 
-void scan_routes(FILE* p_file, route* r){   //reads from a file where each line has the following format: Distance,Track_speed,Power,Gauge,Controls,Station_Start,Station_End
-    char temp[DATA_SIZE];
-    for (int i = 0; i < ROUTE_COUNT; ++i) {
-
-        fscanf(p_file, "%[^,]", temp);
-        r[i].distance = strtol(temp, NULL, 10);  //using strtol here to convert the string we've read into a long integer
-        fscanf(p_file, "%*[,]");
-        fscanf(p_file, "%[^,]", temp);
-        r[i].track_speed = strtol(temp, NULL, 10);
-        fscanf(p_file, "%*[,]");
-        fscanf(p_file, "%[^,]", r[i].power);
-        fscanf(p_file, "%*[,]");
-        fscanf(p_file, "%[^,]", r[i].gauge);        //read until a comma is found
-        fscanf(p_file, "%*[,]");                    //dump the comma
-        fscanf(p_file, "%[^,]", r[i].controls);     //read until the next comma
-        fscanf(p_file, "%*[,]");                    //and so on
-        fscanf(p_file, "%[^,]", r[i].station_start);
-        fscanf(p_file, "%*[,]");
-        fscanf(p_file, "%[^\n]", r[i].station_end); //station_end is the last piece of data. a newline character comes right after it
-        fscanf(p_file, "%*[\n]");           //skip characters until a letter or number is reached, which would be at the next line
-
-        //printf to test if the format is right. delete later
-        printf("%d %d %s %s %s %s %s\n",r[0].distance,r[i].track_speed,r[i].power,r[i].gauge,r[i].controls,r[i].station_start,r[i].station_end);
-    }
-}
-
+// Struct that holds the linked list - aka. the name of route 'i' and a pointer to the next list.
 typedef struct station_list_node{
     char name[DATA_SIZE];
     struct station_list_node* next;
 }station_list_node;
 
-//function that begins a linked list of station names
-station_list_node* create_station_list(const char* first_name){
-    station_list_node* list = malloc(sizeof(station_list_node));   //create the first node using malloc
-    list->next = NULL;                                                  //set the pointer to the next node to NULL since the first node is also the last right now
-    strcpy(list->name ,first_name);                         //add the first station name to the list
-    return list;
-}
+// Function that begins a linked list of station names
+station_list_node* create_station_list(const char* first_name);
 
-//recursive function that tries to add a new name to the station list if it is not already there
-void add_node(station_list_node* node, const char* new_name){
-    if(strcmp(node->name, new_name) != 0){                                       //if the name does not match,
-        if(node->next == NULL){                                                  //we check if we are at the end of the list
-            station_list_node* new_node = malloc(sizeof(station_list_node));//and add a new node if we are
-            new_node->next = NULL;                                               //mostly the same as creating the list, except we set the pointer from the current node to point to the new node instead of NULL
-            strcpy(new_node->name ,new_name);
-            node->next = new_node;
-        } else {                                                                 //if we are not at the end,
-            add_node(node->next, new_name);                                //we try again at the next list node
-        }
-    }
-}
+// Recursive function that tries to add a new name to the station list if it is not already there
+void add_node(station_list_node* node, const char* new_name);
 
-//function to print all names stored in the station list, from first to last
-void print_station_list(station_list_node* node){
-    printf("%s\n", node->name);
-    if(node->next != NULL){
-        print_station_list(node->next);
-    }
-}
+// Function to print all names stored in the station list, from first to last
+void print_station_list(station_list_node* node);
 
-//function to create and fill a linked list of station names with data from the route array. Returns a pointer to the start of the created list
-station_list_node* build_station_list(route* r){
-    station_list_node* list = create_station_list(r[0].station_start);     //use the first data point to begin the list
-    add_node(list, r[0].station_end);
+// Function to create and fill a linked list of station names with data from the route array. Returns a pointer to the start of the created list
+station_list_node* build_station_list(route* r, int* route_count);
 
-    for (int i = 1; i < ROUTE_COUNT; ++i){
-        add_node(list, r[i].station_start); //run add_node with every station_start and station_end value
-        add_node(list, r[i].station_end);
-    }
-    return list;
-}
+// Function to find the length of an existing station list
+int list_length(station_list_node* node);
 
-//function to find the length of an existing station list
-int list_length(station_list_node* node){
-    if(node->next == NULL){
-        return 1;
-    } else {
-        return 1 + list_length(node->next);
-    }
-}
+// Recursive part of search_station_list. the depth variable should always be set to 1 at the beginning of a search
+int search_list_recursive(station_list_node* node, const char* searched_name, int depth);
 
-//recursive part of search_station_list. the depth variable should always be set to 1 at the beginning of a search
-int search_list_recursive(station_list_node* node, const char* searched_name, int depth){
-    if(strcmp(node->name, searched_name) == 0){
-        return 0;
-    } else {
-        if(node->next == NULL){
-            return -(depth);
-        } else {
-            return 1 + search_list_recursive(node->next, searched_name, depth + 1);
-        }
-    }
-}
+// Function that returns an integer based on how far into the list a name is. Returns -1 if searched name is not in the list
+int search_station_list(station_list_node* list, const char* searched_name);
 
-//function that returns an integer based on how far into the list a name is. Returns -1 if searched name is not in the list
-int search_station_list(station_list_node* list, const char* searched_name){
-    return search_list_recursive(list, searched_name, 1);   //runs the recursive search function with depth 1 to make it possible to return -1
-}
+// Returns the name at a given index of a station list starting from the location pointed to by the node variable. Does not support negative index values
+char* index_station_list(station_list_node* node, int index);   // changed 'const int' to 'int'. The function definition is still the same though, as it only affects definitions, not declarations.
 
-//returns the name at a given index of a station list starting from the location pointed to by the node variable. Does not support negative index values
-char* index_station_list(station_list_node* node, const int index){
-    if(index == 0){
-        return node->name;
-    } else {
-        if(node->next == NULL){
-            printf("WARNING: Station list size exceeded by %d", index);
-            return ("INDEX OUT OF BOUNDS");
-        } else {
-            return (index_station_list(node->next, index - 1));
-        }
-    }
-}
-
-void list_test(){   //function to test the linked station list's basic functions
-    station_list_node* list = create_station_list("Test:");
-    char name[DATA_SIZE];
-    for (int i = 0; i < 5; ++i) {
-        scanf("%s", name);
-        add_node(list, name);
-    }
-    print_station_list(list);
-    printf("\nList has a length of %d\n", list_length(list));
-    printf("\nWord \"hello\" is at index %d\n" ,search_station_list(list, "hello"));
-    printf("\nWord at the 3rd index is \"%s\"", index_station_list(list, 3));
-}
+// Function to test the linked station list's basic functions
+void list_test();
 
 #endif //P1_ROUTES_H
